+++ x/metadata/handler_test.go
@@ -22,7 +22,7 @@ import (
 	"github.com/provenance-io/provenance/x/metadata/types/p8e"
 )
 
-type HandlerTestSuite struct {
+type MetadataHandlerTestSuite struct {
 	suite.Suite
 
 	app     *app.App
@@ -38,7 +38,7 @@ type HandlerTestSuite struct {
 	user2Addr sdk.AccAddress
 }
 
-func (s *HandlerTestSuite) SetupTest() {
+func (s *MetadataHandlerTestSuite) SetupTest() {
 	s.app = app.Setup(false)
 	s.ctx = s.app.BaseApp.NewContext(false, tmproto.Header{})
 	s.handler = metadata.NewHandler(s.app.MetadataKeeper)
@@ -54,8 +54,8 @@ func (s *HandlerTestSuite) SetupTest() {
 	s.app.AccountKeeper.SetAccount(s.ctx, s.app.AccountKeeper.NewAccountWithAddress(s.ctx, s.user1Addr))
 }
 
-func TestHandlerTestSuite(t *testing.T) {
-	suite.Run(t, new(HandlerTestSuite))
+func TestMetadataHandlerTestSuite(t *testing.T) {
+	suite.Run(t, new(MetadataHandlerTestSuite))
 }
 
 func createContractSpec(inputSpecs []*p8e.DefinitionSpec, outputSpec p8e.OutputSpec, definitionSpec p8e.DefinitionSpec) p8e.ContractSpec {
@@ -85,7 +85,7 @@ func createDefinitionSpec(name string, classname string, reference p8e.Provenanc
 // TODO: AddScope tests
 // TODO: DeleteScope tests
 
-func (s HandlerTestSuite) TestWriteSession() {
+func (s MetadataHandlerTestSuite) TestWriteSession() {
 	cSpec := types.ContractSpecification{
 		SpecificationId: types.ContractSpecMetadataAddress(uuid.New()),
 		Description:     nil,
@@ -106,9 +106,9 @@ func (s HandlerTestSuite) TestWriteSession() {
 
 	scopeUUID := uuid.New()
 	scope := types.Scope{
-		ScopeId:           types.ScopeMetadataAddress(scopeUUID),
-		SpecificationId:   nil,
-		Owners:            []types.Party{{
+		ScopeId:         types.ScopeMetadataAddress(scopeUUID),
+		SpecificationId: nil,
+		Owners: []types.Party{{
 			Address: s.user1,
 			Role:    types.PartyType_PARTY_TYPE_OWNER,
 		}},
@@ -181,7 +181,7 @@ func (s HandlerTestSuite) TestWriteSession() {
 // TODO: AddRecordSpecification tests
 // TODO: DeleteRecordSpecification tests
 
-func (s HandlerTestSuite) TestAddP8EContractSpec() {
+func (s MetadataHandlerTestSuite) TestAddP8EContractSpec() {
 	validDefSpec := createDefinitionSpec("perform_input_checks", "io.provenance.loan.LoanProtos$PartiesList", p8e.ProvenanceReference{Hash: "Adv+huolGTKofYCR0dw5GHm/R7sUWOwF32XR8r8r9kDy4il5U/LApxOWYHb05jhK4+eY4YzRMRiWcxU3Lx0+Mw=="}, 1)
 	invalidDefSpec := createDefinitionSpec("perform_action", "", p8e.ProvenanceReference{Hash: "Adv+huolGTKofYCR0dw5GHm/R7sUWOwF32XR8r8r9kDy4il5U/LApxOWYHb05jhK4+eY4YzRMRiWcxU3Lx0+Mw=="}, 1)
 
@@ -242,9 +242,10 @@ func ownerPartyList(addresses ...string) []types.Party {
 	return retval
 }
 
-func (s HandlerTestSuite) TestAddAndDeleteScopeDataAccess() {
-	scopeID := types.ScopeMetadataAddress(uuid.New())
+func (s MetadataHandlerTestSuite) TestAddAndDeleteScopeOwners() {
 	scopeSpecID := types.ScopeSpecMetadataAddress(uuid.New())
+	scopeSpec := types.NewScopeSpecification(scopeSpecID, nil, []string{s.user1}, []types.PartyType{types.PartyType_PARTY_TYPE_OWNER}, []types.MetadataAddress{})
+	scopeID := types.ScopeMetadataAddress(uuid.New())
 	scope := types.NewScope(scopeID, scopeSpecID, ownerPartyList(s.user1), []string{s.user1}, "")
 	dneScopeID := types.ScopeMetadataAddress(uuid.New())
 	user3 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address()).String()
@@ -255,6 +256,100 @@ func (s HandlerTestSuite) TestAddAndDeleteScopeDataAccess() {
 		signers  []string
 		errorMsg string
 	}{
+		{
+			"setup test with new scope specification",
+			types.NewMsgWriteScopeSpecificationRequest(*scopeSpec, []string{s.user1}),
+			[]string{s.user1},
+			"",
+		},
+		{
+			"setup test with new scope",
+			types.NewMsgWriteScopeRequest(*scope, []string{s.user1}),
+			[]string{s.user1},
+			"",
+		},
+		{
+			"should fail to ADD owners, msg validate basic failure",
+			types.NewMsgAddScopeOwnerRequest(scopeID, []types.Party{}, []string{s.user1}),
+			[]string{s.user1},
+			"invalid owners: at least one party is required",
+		},
+		{
+			"should fail to ADD owners, can not find scope",
+			types.NewMsgAddScopeOwnerRequest(dneScopeID, []types.Party{{Address: s.user1, Role: types.PartyType_PARTY_TYPE_OWNER}}, []string{s.user1}),
+			[]string{s.user1},
+			fmt.Sprintf("scope not found with id %s", dneScopeID),
+		},
+		{
+			"should fail to ADD owners, validate add failure",
+			types.NewMsgAddScopeOwnerRequest(scopeID, []types.Party{{Address: s.user1, Role: types.PartyType_PARTY_TYPE_OWNER}}, []string{s.user1}),
+			[]string{s.user1},
+			fmt.Sprintf("party already exists with address %s and role %s", s.user1, types.PartyType_PARTY_TYPE_OWNER),
+		},
+		{
+			"should successfully ADD owners",
+			types.NewMsgAddScopeOwnerRequest(scopeID, []types.Party{{Address: s.user2, Role: types.PartyType_PARTY_TYPE_OWNER}}, []string{s.user1}),
+			[]string{s.user1},
+			"",
+		},
+		{
+			"should fail to DELETE owners, msg validate basic failure",
+			types.NewMsgDeleteScopeOwnerRequest(scopeID, []string{}, []string{s.user1, s.user2}),
+			[]string{s.user1},
+			"at least one owner address is required",
+		},
+		{
+			"should fail to DELETE owners, validate add failure",
+			types.NewMsgDeleteScopeOwnerRequest(dneScopeID, []string{s.user1}, []string{s.user1, s.user2}),
+			[]string{s.user1},
+			fmt.Sprintf("scope not found with id %s", dneScopeID),
+		},
+		{
+			"should fail to DELETE owners, validate add failure",
+			types.NewMsgDeleteScopeOwnerRequest(scopeID, []string{user3}, []string{s.user1, s.user2}),
+			[]string{s.user1},
+			fmt.Sprintf("address does not exist in scope owners: %s", user3),
+		},
+		{
+			"should successfully DELETE owners",
+			types.NewMsgDeleteScopeOwnerRequest(scopeID, []string{s.user2}, []string{s.user1, s.user2}),
+			[]string{s.user1},
+			"",
+		},
+	}
+
+	for _, tc := range cases {
+		s.T().Run(tc.name, func(t *testing.T) {
+			_, err := s.handler(s.ctx, tc.msg)
+			if len(tc.errorMsg) > 0 {
+				assert.EqualError(t, err, tc.errorMsg)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
+	}
+}
+
+func (s MetadataHandlerTestSuite) TestAddAndDeleteScopeDataAccess() {
+	scopeSpecID := types.ScopeSpecMetadataAddress(uuid.New())
+	scopeSpec := types.NewScopeSpecification(scopeSpecID, nil, []string{s.user1}, []types.PartyType{types.PartyType_PARTY_TYPE_OWNER}, []types.MetadataAddress{})
+	scopeID := types.ScopeMetadataAddress(uuid.New())
+	scope := types.NewScope(scopeID, scopeSpecID, ownerPartyList(s.user1), []string{s.user1}, "")
+	dneScopeID := types.ScopeMetadataAddress(uuid.New())
+	user3 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address()).String()
+
+	cases := []struct {
+		name     string
+		msg      sdk.Msg
+		signers  []string
+		errorMsg string
+	}{
+		{
+			"setup test with new scope specification",
+			types.NewMsgWriteScopeSpecificationRequest(*scopeSpec, []string{s.user1}),
+			[]string{s.user1},
+			"",
+		},
 		{
 			"setup test with new scope",
 			types.NewMsgWriteScopeRequest(*scope, []string{s.user1}),